Patch to fix hang for memory allocations whose size is bigger than 2^31-1.
Thanks to Bryan Putnam (Purdue) for reporting this.



git-svn-id: https://scm.mvapich.cse.ohio-state.edu/svn/mpi/mvapich2/trunk@1446 09bc9535-d30e-0410-b1f7-d46b20a4725c

+++ src/mpid/osu_ch3/channels/mrail/src/gen2/dreg.c
@@ -956,9 +956,9 @@ dreg_entry *dreg_new_entry(void *buf, int len)
 }
 
 #ifndef DISABLE_PTMALLOC
-void find_and_free_dregs_inside(void *buf, int len)
+void find_and_free_dregs_inside(void *buf, size_t len)
 {
-    int i;
+    unsigned long i;
     unsigned long pagenum_low, pagenum_high;
     unsigned long  npages, begin, end;
     unsigned long user_low_a, user_high_a;
